break out tests to individual unit test files; add conftest and make scope of fixtures module to help with speed
